refatorei o codigo
